// api/src/modules/shared/constants/configuration.enum.ts
export enum Configuration {
  POSTGRESQL_TYPE = 'POSTGRESQL_TYPE',
  POSTGRESQL_HOST = 'POSTGRESQL_HOST',
  POSTGRESQL_PORT = 'POSTGRESQL_PORT',
  POSTGRESQL_DATABASE_NAME = 'POSTGRESQL_DATABASE_NAME',
  POSTGRESQL_USERNAME = 'POSTGRESQL_USERNAME',
  POSTGRESQL_PASSWORD = 'POSTGRESQL_PASSWORD',

  LOGGING = 'LOGGING',
  LOGGER = 'LOGGER',

  ACCESS_SECRET_KEY = 'ACCESS_SECRET_KEY',
  REFRESH_SECRET_KEY = 'REFRESH_SECRET_KEY',
  ACCESS_TOKEN_EXPIRESIN = 'ACCESS_TOKEN_EXPIRESIN',
  REFRESH_TOKEN_EXPIRESIN = 'REFRESH_TOKEN_EXPIRESIN',
  ITEMS_PER_PAGE = 'ITEMS_PER_PAGE',

  BASE_URL = 'BASE_URL',

  SALT = 'SALT',

  STATIC_PATH = 'STATIC_PATH',
  MULTER_DEST = 'MULTER_DEST',
  MAX_FIELD_SIZE_VALUE = 'MAX_FIELD_SIZE_VALUE',
  MAX_FILE_SIZE_NAME = 'MAX_FILE_SIZE_NAME',
  MAX_FILE_SIZE_VALUE = 'MAX_FILE_SIZE_VALUE',
  EXTENSION_NAMES = 'EXTENSION_NAMES',
  EXTENSION_VALUES = 'EXTENSION_VALUES',
}