// api/src/modules/banner/constants/enums/errors.enum.ts
export enum ErrorMessage {
  BANNER_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [banner].',
  LANGUAGE_ID_EMPTY_ERROR = 'Bạn vui lòng chọn [ngôn ngữ].',
  ID_NAN_ERROR = 'Giá trị [id] không hợp lệ.',
  NO_CONTENT = 'Không có dữ liệu hiển thị.',

  OPERATOR_BANNER_ERROR = 'Lưu thông tin [banner] thất bại.',
  BANNER_NOT_FOUND_ERROR = '[Banner] không tồn tại (id: %s)',

  FILE_NOT_FOUND_ERROR = 'Không tìm thấy file.',

  TIME_NAN_ERROR = '[Thời gian] không hợp lệ.',
}